Experimentation Package!
I made an experimentation package so I can have a sort of playground that is kept alongside this project so I can take it anywhere!

Yes, I know you can make a test package alongside your java, but also I am not testing parts of this program if it works and that is what I was taught it meant. ;-;

Oh also now JaredVM_data has Experimentation folder so I can keep track of my progress :heart: